+++ packages/atomic/src/components/common/facets/timeframe-facet-common.tsx
@@ -182,7 +182,7 @@ export class TimeframeFacetCommon {
       return true;
     }
 
-    return !!this.facetForDateRange?.state.values.length;
+    return !!this.valuesToRender.length;
   }
 
   private get numberOfSelectedValues() {