+++ packages/quantic/cypress/integration/facets/numeric-facet/numeric-facet.cypress.ts
@@ -144,6 +144,7 @@ describe('quantic-numeric-facet', () => {
           });
         });
       });
+
       describe('with custom withInput', () => {
         it('should work as expected', () => {
           visitNumericFacetPage({
@@ -422,6 +423,19 @@ describe('quantic-numeric-facet', () => {
           );
         });
       });
+    });
+  });
+
+  describe('with default numeric facet testing accessibility', () => {
+    beforeEach(() => {
+      visitNumericFacetPage(defaultSettings);
+    });
+
+    it('should be accessible through keyboard', () => {
+      Actions.selectFirstNumericFacetValueWithKeyboardTab();
+      Expect.numberOfSelectedCheckboxValues(1);
+      Actions.selectFirstNumericFacetValueWithKeyboardEnter();
+      Expect.numberOfSelectedCheckboxValues(0);
     });
   });
 });